#340 renamed stuff to provider rather than finder, as its not longer finding anything
