+++ src/Ocelot/DependencyInjection/OcelotBuilder.cs
@@ -103,9 +103,9 @@ namespace Ocelot.DependencyInjection
             _services.TryAddSingleton<IUrlPathToUrlTemplateMatcher, RegExUrlMatcher>();
             _services.TryAddSingleton<IPlaceholderNameAndValueFinder, UrlPathPlaceholderNameAndValueFinder>();
             _services.TryAddSingleton<IDownstreamPathPlaceholderReplacer, DownstreamTemplatePathPlaceholderReplacer>();
-            _services.TryAddSingleton<IDownstreamRouteFinder, DownstreamRouteFinder>();
-            _services.TryAddSingleton<IDownstreamRouteFinder, Ocelot.DownstreamRouteFinder.Finder.DownstreamRouteCreator>();
-            _services.TryAddSingleton<IDownstreamRouteFinderFactory, Ocelot.DownstreamRouteFinder.Finder.DownstreamRouteFinderFactory>();
+            _services.TryAddSingleton<IDownstreamRouteProvider, DownstreamRouteProvider>();
+            _services.TryAddSingleton<IDownstreamRouteProvider, Ocelot.DownstreamRouteFinder.Finder.DownstreamRouteCreator>();
+            _services.TryAddSingleton<IDownstreamRouteProviderFactory, Ocelot.DownstreamRouteFinder.Finder.DownstreamRouteProviderFactory>();
             _services.TryAddSingleton<IHttpRequester, HttpClientHttpRequester>();
             _services.TryAddSingleton<IHttpResponder, HttpContextResponder>();
             _services.TryAddSingleton<IErrorsToHttpStatusCodeMapper, ErrorsToHttpStatusCodeMapper>();
