+++ test/Ocelot.UnitTests/DownstreamRouteFinder/DownstreamRouteProviderFactoryTests.cs
@@ -14,19 +14,19 @@ using Xunit;
 
 namespace Ocelot.UnitTests.DownstreamRouteFinder
 {
-    public class DownstreamRouteFinderFactoryTests
+    public class DownstreamRouteProviderFactoryTests
     {
-        private DownstreamRouteFinderFactory _factory;
+        private DownstreamRouteProviderFactory _factory;
 
-        public DownstreamRouteFinderFactoryTests()
+        public DownstreamRouteProviderFactoryTests()
         {
             var services = new ServiceCollection();
             services.AddSingleton<IPlaceholderNameAndValueFinder, UrlPathPlaceholderNameAndValueFinder>();
             services.AddSingleton<IUrlPathToUrlTemplateMatcher, RegExUrlMatcher>();
-            services.AddSingleton<IDownstreamRouteFinder, Ocelot.DownstreamRouteFinder.Finder.DownstreamRouteFinder>();
-            services.AddSingleton<IDownstreamRouteFinder, Ocelot.DownstreamRouteFinder.Finder.DownstreamRouteCreator>();
+            services.AddSingleton<IDownstreamRouteProvider, Ocelot.DownstreamRouteFinder.Finder.DownstreamRouteProvider>();
+            services.AddSingleton<IDownstreamRouteProvider, Ocelot.DownstreamRouteFinder.Finder.DownstreamRouteCreator>();
             var provider = services.BuildServiceProvider();
-            _factory = new DownstreamRouteFinderFactory(provider);
+            _factory = new DownstreamRouteProviderFactory(provider);
         }
 
         //todo - bddfy
@@ -40,7 +40,7 @@ namespace Ocelot.UnitTests.DownstreamRouteFinder
             };
             IInternalConfiguration config = new InternalConfiguration(reRoutes, "", null, "");
             var result = _factory.Get(config);
-            result.ShouldBeOfType<Ocelot.DownstreamRouteFinder.Finder.DownstreamRouteFinder>();
+            result.ShouldBeOfType<Ocelot.DownstreamRouteFinder.Finder.DownstreamRouteProvider>();
         }
 
         [Fact]
@@ -52,7 +52,7 @@ namespace Ocelot.UnitTests.DownstreamRouteFinder
             };
             IInternalConfiguration config = new InternalConfiguration(reRoutes, "", spConfig, "");
             var result = _factory.Get(config);
-            result.ShouldBeOfType<Ocelot.DownstreamRouteFinder.Finder.DownstreamRouteFinder>();
+            result.ShouldBeOfType<Ocelot.DownstreamRouteFinder.Finder.DownstreamRouteProvider>();
         }
 
         [Fact]
